Fixes the ControlPort Java interface library to support the postMessage.
Manages the conversion of the data into the serialized string PMT, but
kept as Strings or ByteBuffers in Java instead of trying to get Java
to understand PMTs.

This requires the latest version of gnuradio.thrift where the message
for the postMessage interface is designed as a binary type instead of
a string.